// src/app/services/openapi-client/model/updateGalleryImageDto.ts
/**
 * ShopErpApi | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateGalleryImageDto { 
    title: string;
    description?: string | null;
    imageUrl: string;
    thumbnailUrl?: string | null;
    altText?: string | null;
    category?: string | null;
    displayOrder?: number;
    isFeatured?: boolean;
    isActive?: boolean;
}